Move tests to a submodule and expose test functions and constants
