+++ zebra-scan/Cargo.toml
@@ -18,6 +18,21 @@ categories = ["cryptography::cryptocurrencies"]
 
 # Production features that activate extra dependencies, or extra features in dependencies
 
+# Test features
+
+proptest-impl = [
+    "proptest",
+    "proptest-derive",
+    "zebra-state/proptest-impl",
+    "zebra-chain/proptest-impl",
+    "bls12_381",
+    "ff",
+    "group",
+    "jubjub",
+    "rand",
+    "zcash_note_encryption",
+]
+
 [dependencies]
 
 color-eyre = "0.6.2"
@@ -37,15 +52,29 @@ zebra-state = { path = "../zebra-state", version = "1.0.0-beta.31", features = [
 
 chrono = { version = "0.4.31", default-features = false, features = ["clock", "std", "serde"] }
 
+# test feature proptest-impl
+proptest = { version = "1.4.0", optional = true }
+proptest-derive = { version = "0.4.0", optional = true }
+
+bls12_381 = { version = "0.8.0", optional = true }
+ff = { version = "0.13.0", optional = true }
+group = { version = "0.13.0", optional = true }
+jubjub = { version = "0.10.0", optional = true }
+rand = { version = "0.8.5", optional = true }
+zcash_note_encryption = { version = "0.4.0", optional = true }
+
 [dev-dependencies]
 
+insta = { version = "1.33.0", features = ["ron", "redactions"] }
+tokio = { version = "1.34.0", features = ["test-util"] }
+
+proptest = "1.4.0"
+proptest-derive = "0.4.0"
 bls12_381 = "0.8.0"
 ff = "0.13.0"
 group = "0.13.0"
 jubjub = "0.10.0"
 rand = "0.8.5"
-tokio = { version = "1.34.0", features = ["test-util"] }
-
 zcash_note_encryption = "0.4.0"
 
 zebra-state = { path = "../zebra-state", version = "1.0.0-beta.31", features = ["proptest-impl"] }
