+++ zebra-scan/src/tests.rs
@@ -12,17 +12,14 @@ use ff::{Field, PrimeField};
 use group::GroupEncoding;
 use rand::{rngs::OsRng, thread_rng, RngCore};
 
-use zcash_client_backend::{
-    encoding::decode_extended_full_viewing_key,
-    proto::compact_formats::{
-        ChainMetadata, CompactBlock, CompactSaplingOutput, CompactSaplingSpend, CompactTx,
-    },
+use zcash_client_backend::proto::compact_formats::{
+    ChainMetadata, CompactBlock, CompactSaplingOutput, CompactSaplingSpend, CompactTx,
 };
 use zcash_note_encryption::Domain;
 use zcash_primitives::{
     block::BlockHash,
     consensus::BlockHeight,
-    constants::{mainnet::HRP_SAPLING_EXTENDED_FULL_VIEWING_KEY, SPENDING_KEY_GENERATOR},
+    constants::SPENDING_KEY_GENERATOR,
     memo::MemoBytes,
     sapling::{
         note_encryption::{sapling_note_encryption, SaplingDomain},
@@ -30,200 +27,29 @@ use zcash_primitives::{
         value::NoteValue,
         Note, Nullifier,
     },
-    zip32::{DiversifiableFullViewingKey, ExtendedSpendingKey},
+    zip32::DiversifiableFullViewingKey,
 };
 
 use zebra_chain::{
     amount::{Amount, NegativeAllowed},
     block::{self, merkle, Block, Header, Height},
-    chain_tip::ChainTip,
     fmt::HexDebug,
-    parameters::Network,
     primitives::{redjubjub, Groth16Proof},
     sapling::{self, PerSpendAnchor, Spend, TransferData},
-    serialization::{AtLeastOne, ZcashDeserializeInto},
+    serialization::AtLeastOne,
     transaction::{LockTime, Transaction},
     transparent::{CoinbaseData, Input},
     work::{difficulty::CompactDifficulty, equihash::Solution},
 };
-use zebra_state::{SaplingScannedResult, TransactionIndex};
 
-use crate::{
-    config::Config,
-    scan::{block_to_compact, scan_block},
-};
+#[cfg(test)]
+mod vectors;
 
-/// This test:
-/// - Creates a viewing key and a fake block containing a Sapling output decryptable by the key.
-/// - Scans the block.
-/// - Checks that the result contains the txid of the tx containing the Sapling output.
-#[tokio::test]
-async fn scanning_from_fake_generated_blocks() -> Result<()> {
-    let extsk = ExtendedSpendingKey::master(&[]);
-    let dfvk: DiversifiableFullViewingKey = extsk.to_diversifiable_full_viewing_key();
-    let nf = Nullifier([7; 32]);
+/// The extended Sapling viewing key of [ZECpages](https://zecpages.com/boardinfo)
+pub const ZECPAGES_SAPLING_VIEWING_KEY: &str = "zxviews1q0duytgcqqqqpqre26wkl45gvwwwd706xw608hucmvfalr759ejwf7qshjf5r9aa7323zulvz6plhttp5mltqcgs9t039cx2d09mgq05ts63n8u35hyv6h9nc9ctqqtue2u7cer2mqegunuulq2luhq3ywjcz35yyljewa4mgkgjzyfwh6fr6jd0dzd44ghk0nxdv2hnv4j5nxfwv24rwdmgllhe0p8568sgqt9ckt02v2kxf5ahtql6s0ltjpkckw8gtymxtxuu9gcr0swvz";
 
-    let (block, sapling_tree_size) = fake_block(1u32.into(), nf, &dfvk, 1, true, Some(0));
-
-    assert_eq!(block.transactions.len(), 4);
-
-    let res = scan_block(Network::Mainnet, &block, sapling_tree_size, &[&dfvk]).unwrap();
-
-    // The response should have one transaction relevant to the key we provided.
-    assert_eq!(res.transactions().len(), 1);
-
-    // Check that the original block contains the txid in the scanning result.
-    assert!(block
-        .transactions
-        .iter()
-        .map(|tx| tx.hash().bytes_in_display_order())
-        .any(|txid| &txid == res.transactions()[0].txid.as_ref()));
-
-    // Check that the txid in the scanning result matches the third tx in the original block.
-    assert_eq!(
-        res.transactions()[0].txid.as_ref(),
-        &block.transactions[2].hash().bytes_in_display_order()
-    );
-
-    // The block hash of the response should be the same as the one provided.
-    assert_eq!(res.block_hash().0, block.hash().0);
-
-    Ok(())
-}
-
-/// Scan a populated state for the ZECpages viewing key.
-/// This test is very similar to `scanning_from_populated_zebra_state` but with the ZECpages key.
-/// There are no zechub transactions in the test data so we should get empty related transactions.
-#[tokio::test]
-async fn scanning_zecpages_from_populated_zebra_state() -> Result<()> {
-    /// The extended Sapling viewing key of [ZECpages](https://zecpages.com/boardinfo)
-    const ZECPAGES_VIEWING_KEY: &str = "zxviews1q0duytgcqqqqpqre26wkl45gvwwwd706xw608hucmvfalr759ejwf7qshjf5r9aa7323zulvz6plhttp5mltqcgs9t039cx2d09mgq05ts63n8u35hyv6h9nc9ctqqtue2u7cer2mqegunuulq2luhq3ywjcz35yyljewa4mgkgjzyfwh6fr6jd0dzd44ghk0nxdv2hnv4j5nxfwv24rwdmgllhe0p8568sgqt9ckt02v2kxf5ahtql6s0ltjpkckw8gtymxtxuu9gcr0swvz";
-
-    // Parse the key from ZECpages
-    let efvk = decode_extended_full_viewing_key(
-        HRP_SAPLING_EXTENDED_FULL_VIEWING_KEY,
-        ZECPAGES_VIEWING_KEY,
-    )
-    .unwrap();
-
-    // Build a vector of viewing keys `vks` to scan for.
-    let fvk = efvk.fvk;
-    let ivk = fvk.vk.ivk();
-    let ivks = vec![ivk];
-
-    let network = Network::Mainnet;
-
-    // Create a continuous chain of mainnet blocks from genesis
-    let blocks: Vec<Arc<Block>> = zebra_test::vectors::CONTINUOUS_MAINNET_BLOCKS
-        .iter()
-        .map(|(_height, block_bytes)| block_bytes.zcash_deserialize_into().unwrap())
-        .collect();
-
-    // Create a populated state service.
-    let (_state_service, read_only_state_service, latest_chain_tip, _chain_tip_change) =
-        zebra_state::populated_state(blocks.clone(), network).await;
-
-    let db = read_only_state_service.db();
-
-    // use the tip as starting height
-    let mut height = latest_chain_tip.best_tip_height().unwrap();
-
-    let mut transactions_found = 0;
-    let mut transactions_scanned = 0;
-    let mut blocks_scanned = 0;
-    while let Some(block) = db.block(height.into()) {
-        // We use a dummy size of the Sapling note commitment tree. We can't set the size to zero
-        // because the underlying scanning function would return
-        // `zcash_client_backeng::scanning::ScanError::TreeSizeUnknown`.
-        let sapling_commitment_tree_size = 1;
-
-        let orchard_commitment_tree_size = 0;
-
-        let chain_metadata = ChainMetadata {
-            sapling_commitment_tree_size,
-            orchard_commitment_tree_size,
-        };
-
-        let compact_block = block_to_compact(&block, chain_metadata);
-
-        let res = scan_block(network, &block, sapling_commitment_tree_size, &ivks)
-            .expect("scanning block for the ZECpages viewing key should work");
-
-        transactions_found += res.transactions().len();
-        transactions_scanned += compact_block.vtx.len();
-        blocks_scanned += 1;
-
-        // scan backwards
-        if height.is_min() {
-            break;
-        }
-        height = height.previous()?;
-    }
-
-    // make sure all blocks and transactions were scanned
-    assert_eq!(blocks_scanned, 11);
-    assert_eq!(transactions_scanned, 11);
-
-    // no relevant transactions should be found
-    assert_eq!(transactions_found, 0);
-
-    Ok(())
-}
-
-/// Creates a viewing key and a fake block containing a Sapling output decryptable by the key, scans
-/// the block using the key, and adds the results to the database.
-///
-/// The purpose of this test is to check if our database and our scanning code are compatible.
-#[test]
-#[allow(deprecated)]
-fn scanning_fake_blocks_store_key_and_results() -> Result<()> {
-    // Generate a key
-    let extsk = ExtendedSpendingKey::master(&[]);
-    // TODO: find out how to do it with `to_diversifiable_full_viewing_key` as `to_extended_full_viewing_key` is deprecated.
-    let extfvk = extsk.to_extended_full_viewing_key();
-    let dfvk: DiversifiableFullViewingKey = extsk.to_diversifiable_full_viewing_key();
-    let key_to_be_stored =
-        zcash_client_backend::encoding::encode_extended_full_viewing_key("zxviews", &extfvk);
-
-    // Create a database
-    let mut s = crate::storage::Storage::new(&Config::ephemeral(), Network::Mainnet);
-
-    // Insert the generated key to the database
-    s.add_sapling_key(&key_to_be_stored, None);
-
-    // Check key was added
-    assert_eq!(s.sapling_keys().len(), 1);
-    assert_eq!(
-        s.sapling_keys().get(&key_to_be_stored),
-        Some(&s.min_sapling_birthday_height())
-    );
-
-    let nf = Nullifier([7; 32]);
-
-    let (block, sapling_tree_size) = fake_block(1u32.into(), nf, &dfvk, 1, true, Some(0));
-
-    let result = scan_block(Network::Mainnet, &block, sapling_tree_size, &[&dfvk]).unwrap();
-
-    // The response should have one transaction relevant to the key we provided.
-    assert_eq!(result.transactions().len(), 1);
-
-    let result = SaplingScannedResult::from(result.transactions()[0].txid.as_ref());
-
-    // Add result to database
-    s.add_sapling_results(
-        key_to_be_stored.clone(),
-        Height(1),
-        [(TransactionIndex::from_usize(0), result)].into(),
-    );
-
-    // Check the result was added
-    assert_eq!(
-        s.sapling_results(&key_to_be_stored).get(&Height(1)),
-        Some(&vec![result])
-    );
-
-    Ok(())
-}
+/// A fake viewing key in an incorrect format.
+pub const FAKE_SAPLING_VIEWING_KEY: &str = "zxviewsfake";
 
 /// Generates a fake block containing a Sapling output decryptable by `dfvk`.
 ///
@@ -232,7 +58,7 @@ fn scanning_fake_blocks_store_key_and_results() -> Result<()> {
 /// 2. a V4 tx containing a random Sapling output,
 /// 3. a V4 tx containing a Sapling output decryptable by `dfvk`,
 /// 4. depending on the value of `tx_after`, another V4 tx containing a random Sapling output.
-fn fake_block(
+pub fn fake_block(
     height: BlockHeight,
     nf: Nullifier,
     dfvk: &DiversifiableFullViewingKey,
@@ -305,7 +131,7 @@ fn fake_block(
 // be a number for easier conversion:
 // https://github.com/zcash/librustzcash/blob/zcash_primitives-0.13.0/zcash_client_backend/src/scanning.rs#L635
 // We need to copy because this is a test private function upstream.
-fn fake_compact_block(
+pub fn fake_compact_block(
     height: BlockHeight,
     prev_hash: BlockHash,
     nf: Nullifier,
@@ -394,7 +220,7 @@ fn fake_compact_block(
 // This is an exact copy of `zcash_client_backend::scanning::random_compact_tx`:
 // https://github.com/zcash/librustzcash/blob/zcash_primitives-0.13.0/zcash_client_backend/src/scanning.rs#L597
 // We need to copy because this is a test private function upstream.
-fn random_compact_tx(mut rng: impl RngCore) -> CompactTx {
+pub fn random_compact_tx(mut rng: impl RngCore) -> CompactTx {
     let fake_nf = {
         let mut nf = vec![0; 32];
         rng.fill_bytes(&mut nf);
@@ -427,7 +253,7 @@ fn random_compact_tx(mut rng: impl RngCore) -> CompactTx {
 }
 
 /// Converts [`CompactTx`] to [`Transaction::V4`].
-fn compact_to_v4(tx: &CompactTx) -> Result<Transaction> {
+pub fn compact_to_v4(tx: &CompactTx) -> Result<Transaction> {
     let sk = redjubjub::SigningKey::<redjubjub::SpendAuth>::new(thread_rng());
     let vk = redjubjub::VerificationKey::from(&sk);
     let dummy_rk = sapling::keys::ValidatingKey::try_from(vk)
